+++ src/config.service.ts
@@ -16,39 +16,38 @@ export interface IPopupOptions {
 export interface IOauth1Options {
     url?: string;
     name?: string;
-    popupOptions?: IPopupOptions;
     redirectUri?: string;
+    popupOptions?: IPopupOptions;
     authorizationEndpoint?: string;
-    oauthType?: string;
-    exchangeForToken?: boolean | ((options: {code?: string, state?: string}, userData?: any) => Observable<Response>);
+    oauthType?: '1.0';
     method?: string;
 }
 
-export interface IOauth2Options extends IOauth1Options {
-    state?: string| (() => string);
+export interface IOauth2Options {
+    url?: string;
+    name?: string;
+    redirectUri?: string;
+    popupOptions?: IPopupOptions;
+    authorizationEndpoint?: string;
+    oauthType?: '2.0';
+    method?: string;
     responseType?: string;
-    responseParams?: {
-        code?: string;
-        clientId?: string;
-        redirectUri?: string;
-    };
     clientId?: string;
+    additionalUrlParams?: {
+        [paramName: string]: string | (() => string) | null | undefined;
+    };
     scopeDelimiter?: string;
-    scopePrefix?: string;
-
-    requiredUrlParams?: string[];
-    optionalUrlParams?: string[];
-    scope?: string[]|string;
-    display?: string;
+    scope?: string[];
+    state?: string | (() => string);
 }
 
 export interface IProviders {
-    [provider: string]: IOauth2Options;
+    [provider: string]: IOauth2Options | IOauth1Options;
 }
 
-export abstract class CustomConfig {
-    tokenRoot: string;
-    cordova: boolean;
+export interface IConfigOptions {
+    tokenRoot: string | null;
+    cordova: boolean | null;
     baseUrl: string;
     loginUrl: string;
     signupUrl: string;
@@ -61,77 +60,115 @@ export abstract class CustomConfig {
     storageType: 'localStorage' | 'sessionStorage' | 'cookie' | 'sessionCookie';
     providers: IProviders;
     withCredentials: boolean;
-    autoRefreshToken: boolean;
-    refreshUrl: string;
-    refreshBeforeExpiration: number;
-    tryTokenRefreshIfUnauthorized: boolean;
-    resolveToken: (response: Response) => string;
+    resolveToken: (response: any, config: IConfigOptions) => string;
 }
 
-@Injectable()
-export class ConfigService {
-    withCredentials = false;
-    tokenRoot = null;
-    baseUrl = '/';
-    loginUrl = '/auth/login';
-    signupUrl = '/auth/signup';
-    unlinkUrl = '/auth/unlink/';
-    refreshUrl = '/auth/refresh';
-    tokenName = 'token';
-    tokenSeparator = '_';
-    tokenPrefix = 'ng2-ui-auth';
-    authHeader = 'Authorization';
-    authToken = 'Bearer';
-    storageType: 'localStorage' | 'sessionStorage' | 'cookie' | 'sessionCookie' = 'localStorage';
-    autoRefreshToken = false;
-    refreshBeforeExpiration = 600000; // 10 minutes
-    tryTokenRefreshIfUnauthorized = false;
-    cordova = this.isCordovaApp();
-    providers: IProviders = {
+export interface IPartialConfigOptions { // = Partial<IConfigOptions
+    tokenRoot?: string | null;
+    cordova?: boolean | null;
+    baseUrl?: string;
+    loginUrl?: string;
+    signupUrl?: string;
+    unlinkUrl?: string;
+    tokenName?: string;
+    tokenSeparator?: string;
+    tokenPrefix?: string;
+    authToken?: string;
+    authHeader?: string;
+    storageType?: 'localStorage' | 'sessionStorage' | 'cookie' | 'sessionCookie';
+    providers?: IProviders;
+    withCredentials?: boolean;
+    resolveToken?: (response: any, config: IConfigOptions) => string;
+}
+
+export const defaultOptions: IConfigOptions = {
+    withCredentials: false,
+    tokenRoot: null,
+    baseUrl: '/',
+    loginUrl: '/auth/login',
+    signupUrl: '/auth/signup',
+    unlinkUrl: '/auth/unlink/',
+    tokenName: 'token',
+    tokenSeparator: '_',
+    tokenPrefix: 'ng2-ui-auth',
+    authHeader: 'Authorization',
+    authToken: 'Bearer',
+    storageType: 'localStorage',
+    cordova: null,
+    resolveToken: (response: any, config: IConfigOptions) => {
+        const accessToken: string | { [key: string]: string } | null | undefined = response &&
+            (response.access_token || response.token || response.data);
+        if (!accessToken) {
+            // console.warn('No token found');
+            return null;
+        }
+        if (typeof accessToken === 'string') {
+            return accessToken;
+        }
+        if (typeof accessToken !== 'object') {
+            // console.warn('No token found');
+            return null;
+        }
+        const tokenRootData = config.tokenRoot && config.tokenRoot
+            .split('.')
+            .reduce(
+            (o: any, x: any) => {
+                return o[x];
+            },
+            accessToken);
+        const token = tokenRootData ? tokenRootData[config.tokenName] : accessToken[config.tokenName];
+        if (token) {
+            return token;
+        }
+        // const tokenPath = this.tokenRoot ? this.tokenRoot + '.' + this.tokenName : this.tokenName;
+        // console.warn('Expecting a token named "' + tokenPath);
+        return null;
+    },
+    providers: {
         facebook: {
             name: 'facebook',
             url: '/auth/facebook',
             authorizationEndpoint: 'https://www.facebook.com/v2.5/dialog/oauth',
-            redirectUri: this.getHttpHost('/'),
-            requiredUrlParams: ['display', 'scope'],
+            additionalUrlParams: {
+                display: 'popup',
+            },
             scope: ['email'],
             scopeDelimiter: ',',
-            display: 'popup',
             oauthType: '2.0',
-            popupOptions: {width: 580, height: 400},
+            popupOptions: { width: 580, height: 400 },
         },
         google: {
             name: 'google',
             url: '/auth/google',
             authorizationEndpoint: 'https://accounts.google.com/o/oauth2/auth',
-            redirectUri: this.getHttpHost(),
-            requiredUrlParams: ['scope'],
-            optionalUrlParams: ['display', 'state', 'prompt', 'login_hint', 'access_type', 'include_granted_scopes', 'openid.realm', 'hd'],
-            scope: ['profile', 'email'],
-            scopePrefix: 'openid',
+            additionalUrlParams: {
+                'display': 'popup',
+                'prompt': undefined,
+                'login_hint': undefined,
+                'access_type': undefined,
+                'include_granted_scopes': undefined,
+                'openid.realm': undefined,
+                'hd': undefined,
+            },
+            scope: ['openid', 'profile', 'email'],
             scopeDelimiter: ' ',
-            display: 'popup',
             oauthType: '2.0',
-            popupOptions: {width: 452, height: 633},
+            popupOptions: { width: 452, height: 633 },
             state: () => encodeURIComponent(Math.random().toString(36).substr(2)),
         },
         github: {
             name: 'github',
             url: '/auth/github',
             authorizationEndpoint: 'https://github.com/login/oauth/authorize',
-            redirectUri: this.getHttpHost(),
-            optionalUrlParams: ['scope'],
             scope: ['user:email'],
             scopeDelimiter: ' ',
             oauthType: '2.0',
-            popupOptions: {width: 1020, height: 618},
+            popupOptions: { width: 1020, height: 618 },
         },
         instagram: {
             name: 'instagram',
             url: '/auth/instagram',
             authorizationEndpoint: 'https://api.instagram.com/oauth/authorize',
-            redirectUri: this.getHttpHost(),
-            requiredUrlParams: ['scope'],
             scope: ['basic'],
             scopeDelimiter: '+',
             oauthType: '2.0',
@@ -140,139 +177,92 @@ export class ConfigService {
             name: 'linkedin',
             url: '/auth/linkedin',
             authorizationEndpoint: 'https://www.linkedin.com/uas/oauth2/authorization',
-            redirectUri: this.getHttpHost(),
-            requiredUrlParams: ['state'],
             scope: ['r_emailaddress'],
             scopeDelimiter: ' ',
-            state: 'STATE',
             oauthType: '2.0',
-            popupOptions: {width: 527, height: 582},
+            popupOptions: { width: 527, height: 582 },
+            state: 'STATE',
         },
         twitter: {
             name: 'twitter',
             url: '/auth/twitter',
             authorizationEndpoint: 'https://api.twitter.com/oauth/authenticate',
-            redirectUri: this.getHttpHost(),
             oauthType: '1.0',
-            popupOptions: {width: 495, height: 645},
+            popupOptions: { width: 495, height: 645 },
         },
         twitch: {
             name: 'twitch',
             url: '/auth/twitch',
             authorizationEndpoint: 'https://api.twitch.tv/kraken/oauth2/authorize',
-            redirectUri: this.getHttpHost(),
-            requiredUrlParams: ['scope'],
             scope: ['user_read'],
             scopeDelimiter: ' ',
-            display: 'popup',
+            additionalUrlParams: {
+                display: 'popup',
+            },
             oauthType: '2.0',
-            popupOptions: {width: 500, height: 560},
+            popupOptions: { width: 500, height: 560 },
         },
         live: {
             name: 'live',
             url: '/auth/live',
             authorizationEndpoint: 'https://login.live.com/oauth20_authorize.srf',
-            redirectUri: this.getHttpHost(),
-            requiredUrlParams: ['display', 'scope'],
+            additionalUrlParams: {
+                display: 'popup',
+            },
             scope: ['wl.emails'],
             scopeDelimiter: ' ',
-            display: 'popup',
             oauthType: '2.0',
-            popupOptions: {width: 500, height: 560},
+            popupOptions: { width: 500, height: 560 },
         },
         yahoo: {
             name: 'yahoo',
             url: '/auth/yahoo',
             authorizationEndpoint: 'https://api.login.yahoo.com/oauth2/request_auth',
-            redirectUri: this.getHttpHost(),
             scope: [],
             scopeDelimiter: ',',
             oauthType: '2.0',
-            popupOptions: {width: 559, height: 519},
+            popupOptions: { width: 559, height: 519 },
         },
         bitbucket: {
             name: 'bitbucket',
             url: '/auth/bitbucket',
             authorizationEndpoint: 'https://bitbucket.org/site/oauth2/authorize',
-            redirectUri: this.getHttpHost('/'),
-            requiredUrlParams: ['scope'],
             scope: ['email'],
             scopeDelimiter: ',',
             oauthType: '2.0',
-            popupOptions: {width: 1028, height: 529},
+            popupOptions: { width: 1028, height: 529 },
         },
         spotify: {
             name: 'spotify',
             url: '/auth/spotify',
             authorizationEndpoint: 'https://accounts.spotify.com/authorize',
-            redirectUri: this.getHttpHost(),
-            optionalUrlParams: ['state'],
-            requiredUrlParams: ['scope'],
-            scope: ['user-read-email'],
-            scopePrefix: '',
+            scope: ['', 'user-read-email'],
             scopeDelimiter: ',',
             oauthType: '2.0',
             popupOptions: { width: 500, height: 530 },
             state: () => encodeURIComponent(Math.random().toString(36).substr(2)),
         },
-    };
+    },
+};
 
-    constructor(config?: CustomConfig) {
-        Object.keys(config).forEach((key) => {
-            if (typeof config[key] === 'undefined') {
-                return;
-            }
-            if (key !== 'providers') {
-                this[key] = config[key];
-            } else {
-                Object.keys(config[key]).map(provider => {
-                    this.providers[provider] = Object.assign(
-                        this.providers[provider] || {},
-                        config.providers[provider],
-                    );
-                });
-            }
-        });
-    }
+@Injectable()
+export class ConfigService {
+    public options: IConfigOptions;
 
-    resolveToken = (response: Response|Object) => {
-        let tokenObj = response;
-        if (response instanceof Response) {
-            tokenObj = response.json();
-        }
-        const accessToken: string | Object | null | undefined = tokenObj &&
-            (tokenObj['access_token'] || tokenObj['token'] || tokenObj['data']);
-        if (!accessToken) {
-            console.warn('No token found');
-            return null;
-        }
-        if (typeof accessToken === 'string') {
-            return accessToken;
-        }
-        if (typeof accessToken !== 'object') {
-            console.warn('No token found');
-            return null;
-        }
-        const tokenRootData = this.tokenRoot &&
-            this.tokenRoot.split('.').reduce(
-                (o, x) => {
-                    return o[x];
-                },
-                accessToken);
-        const token = tokenRootData ? tokenRootData[this.tokenName] : accessToken[this.tokenName];
-        if (token) {
-            return token;
-        }
-        let tokenPath = this.tokenRoot ? this.tokenRoot + '.' + this.tokenName : this.tokenName;
-        console.warn('Expecting a token named "' + tokenPath);
-        return null;
-    }
-
-    getHttpHost(path = '') {
-        return window.location.origin + path;
-    }
-
-    isCordovaApp() {
-        return !!window['cordova'];
+    constructor(options: IPartialConfigOptions) {
+        this.options = {
+            ...defaultOptions,
+            ...options,
+            providers: {
+                ...options.providers,
+                ...Object
+                    .keys(defaultOptions.providers)
+                    .concat(Object.keys(options.providers || {}))
+                    .map((key) => options.providers && options.providers[key]
+                        ? { [key]: { ...defaultOptions.providers[key], ...options.providers[key] } }
+                        : { [key]: defaultOptions.providers[key] })
+                    .reduce((acc, next) => ({ ...acc, ...next }), {}),
+            },
+        } as IConfigOptions;
     }
 }