+++ src/interceptor.service.ts
@@ -1,7 +1,7 @@
 import { switchMapTo } from 'rxjs/operators';
 import { ConfigService } from './config.service';
 import { SharedService } from './shared.service';
-import { Injectable } from '@angular/core';
+import { Injectable, Injector } from '@angular/core';
 import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest, HttpClient, HttpErrorResponse, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { tap, catchError } from 'rxjs/operators';
@@ -9,46 +9,18 @@ import { tap, catchError } from 'rxjs/operators';
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
     constructor(
-        private _http: HttpClient,
-        private _shared: SharedService,
-        private _config: ConfigService,
+        private shared: SharedService,
+        private config: ConfigService,
     ) { }
+
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        const { authHeader, refreshBeforeExpiration, autoRefreshToken } = this._config;
-        const token = this._shared.getToken();
-        const isAuthenticated = this._shared.isAuthenticated;
-        const exp = this._shared.getExpirationDate();
-        const newReq = isAuthenticated
-            ? req.clone({ setHeaders: { [authHeader]: this.getAuthHeaderString() } })
+        const { authHeader, authToken } = this.config.options;
+        const token = this.shared.getToken();
+        const isAuthenticated = this.shared.isAuthenticated;
+        const newReq = isAuthenticated && !req.headers.has(authHeader)
+            ? req.clone({ setHeaders: { [authHeader]: `${authToken} ${token}` } })
             : req;
-
-        if (token &&
-            (!exp || exp.getTime() + refreshBeforeExpiration > Date.now()) &&
-            autoRefreshToken) {
-            return this.refreshToken().pipe(switchMapTo(next.handle(newReq)));
-        }
-        if (this._config.tryTokenRefreshIfUnauthorized) {
-            return next.handle(newReq)
-                .pipe(catchError((response: HttpErrorResponse) => {
-                    if (response.status === 401) {
-                        return this.refreshToken().pipe(switchMapTo(next.handle(newReq)));
-                    }
-                    throw response;
-                }));
-        }
         return next.handle(newReq);
-    }
-
-    refreshToken(): Observable<any> {
-        return this._http
-            .get(this._config.refreshUrl, {
-                headers: new HttpHeaders({ [this._config.authHeader]: this.getAuthHeaderString() }),
-            })
-            .pipe(tap(this._shared.setToken));
-    }
-
-    getAuthHeaderString() {
-        return this._config.authToken + ' ' + this._shared.getToken();
     }
 
 }