+++ src/local.service.ts
@@ -17,12 +17,12 @@ export class LocalService {
         private shared: SharedService,
         private config: ConfigService) { }
 
-    login<T extends string | object>(user: string | object, url?: string): Observable<T> {
-        return this.http.post<T>(url || joinUrl(this.config.baseUrl, this.config.loginUrl), user)
-            .pipe(tap(data => this.shared.setToken(data)));
+    public login<T extends string | object>(user: string | object, url?: string): Observable<T> {
+        return this.http.post<T>(url || joinUrl(this.config.options.baseUrl, this.config.options.loginUrl), user)
+            .pipe(tap((data) => this.shared.setToken(data)));
     }
 
-    signup<T = any>(user: string | Object, url?: string): Observable<T> {
-        return this.http.post<T>(url || joinUrl(this.config.baseUrl, this.config.loginUrl), user);
+    public signup<T = any>(user: string | object, url?: string): Observable<T> {
+        return this.http.post<T>(url || joinUrl(this.config.options.baseUrl, this.config.options.loginUrl), user);
     }
 }