+++ src/oauth1.service.ts
@@ -1,6 +1,7 @@
+import { IOauthService } from './oauth-service.interface';
 import { Injectable } from '@angular/core';
 import { PopupService } from './popup.service';
-import { joinUrl } from './utils';
+import { joinUrl, buildQueryString } from './utils';
 import { ConfigService, IOauth1Options } from './config.service';
 import { Observable } from 'rxjs/Observable';
 import { switchMap } from 'rxjs/operators';
@@ -10,68 +11,34 @@ import { HttpClient } from '@angular/common/http';
  * Created by Ron on 17/12/2015.
  */
 @Injectable()
-export class Oauth1Service {
-    private static base: IOauth1Options = {
-        url: null,
-        name: null,
-        popupOptions: null,
-        redirectUri: null,
-        authorizationEndpoint: null,
-    };
-    private defaults: IOauth1Options;
+export class Oauth1Service implements IOauthService {
 
-    constructor(private http: HttpClient, private popup: PopupService, private config: ConfigService) {
+    constructor(
+        private http: HttpClient,
+        private popup: PopupService,
+        private config: ConfigService,
+    ) { }
+
+    open<T extends object | string = any>(oauthOptions: IOauth1Options, userData: object): Observable<T> {
+        const serverUrl = this.config.options.baseUrl
+            ? joinUrl(this.config.options.baseUrl, oauthOptions.url)
+            : oauthOptions.url;
+
+        return this.http.post<object>(serverUrl, oauthOptions).pipe(
+            switchMap((authorizationData) => this.popup.open(
+                [oauthOptions.authorizationEndpoint, buildQueryString(authorizationData)].join('?'),
+                oauthOptions,
+                this.config.options.cordova,
+            ), (authorizationData, oauthData) => ({ authorizationData, oauthData })),
+            switchMap(({ authorizationData, oauthData }) => this.exchangeForToken<T>(oauthOptions, authorizationData, oauthData, userData)),
+        );
     }
 
-    open<T>(options?: IOauth1Options, userData?: any): Observable<T> {
-        this.defaults = { ...Oauth1Service.base, ...options };
-        let popupWindow;
-        let serverUrl = this.config.baseUrl ? joinUrl(this.config.baseUrl, this.defaults.url) : this.defaults.url;
-
-        if (!this.config.cordova) {
-            popupWindow = this.popup.open('', this.defaults.name, this.defaults.popupOptions/*, this.defaults.redirectUri*/);
-        }
-
-        return this.http
-            .post<T>(serverUrl, JSON.stringify(this.defaults))
-            .pipe(
-            switchMap(response => {
-                if (this.config.cordova) {
-                    popupWindow = this.popup.open(
-                        [this.defaults.authorizationEndpoint, this.buildQueryString(response)].join('?'),
-                        this.defaults.name,
-                        this.defaults.popupOptions);
-                } else {
-                    popupWindow.popupWindow.location =
-                        [this.defaults.authorizationEndpoint, this.buildQueryString(response)].join('?');
-                }
-
-                return this.config.cordova ? popupWindow.eventListener(this.defaults.redirectUri) : popupWindow.pollPopup();
-            }),
-            switchMap(response => {
-                let exchangeForToken: any = options.exchangeForToken;
-                if (typeof exchangeForToken !== 'function') {
-                    exchangeForToken = this.exchangeForToken.bind(this);
-                }
-                return exchangeForToken(response, userData);
-            }),
-            );
-    }
-
-    private exchangeForToken(oauthData, userData?: any) {
-        let data = { ...this.defaults, ...oauthData, ...userData };
-        let exchangeForTokenUrl = this.config.baseUrl ? joinUrl(this.config.baseUrl, this.defaults.url) : this.defaults.url;
-        return this.defaults.method
-            ? this.http.request(this.defaults.method, exchangeForTokenUrl, {
-                body: data,
-                withCredentials: this.config.withCredentials,
-            })
-            : this.http.post(exchangeForTokenUrl, data, { withCredentials: this.config.withCredentials });
-    }
-
-    private buildQueryString(obj: Object) {
-        return Object.keys(obj).map((key) => {
-            return encodeURIComponent(key) + '=' + encodeURIComponent(obj[key]);
-        }).join('&');
+    private exchangeForToken<T>(oauthOptions: IOauth1Options, authorizationData: object, oauthData: object, userData: object) {
+        const body = { authorizationData, oauthData, userData };
+        const { withCredentials, baseUrl } = this.config.options;
+        const { method = 'POST', url } = oauthOptions;
+        const exchangeForTokenUrl = baseUrl ? joinUrl(baseUrl, url) : url;
+        return this.http.request<T>(method, exchangeForTokenUrl, { body, withCredentials })
     }
 }