+++ src/oauth2.service.ts
@@ -1,6 +1,7 @@
+import { IOauthService } from './oauth-service.interface';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { camelCase, joinUrl, deepMerge } from './utils';
+import { joinUrl, deepMerge, buildQueryString } from './utils';
 import { ConfigService, IOauth2Options } from './config.service';
 import { PopupService } from './popup.service';
 import { StorageService } from './storage.service';
@@ -14,120 +15,76 @@ import { of } from 'rxjs/observable/of';
 
 
 @Injectable()
-export class Oauth2Service {
-    private static base: IOauth2Options & { defaultUrlParams: string[] } = {
-        defaultUrlParams: ['response_type', 'client_id', 'redirect_uri'],
-        responseType: 'code',
-        responseParams: {
-            code: 'code',
-            clientId: 'clientId',
-            redirectUri: 'redirectUri',
-        },
-    };
-
-    private defaults: IOauth2Options & { defaultUrlParams: string[] };
+export class Oauth2Service implements IOauthService {
 
     constructor(
         private http: HttpClient,
         private popup: PopupService,
-        private storage: StorageService,
         private config: ConfigService) {
     }
 
-    open<T>(options?: IOauth2Options, userData?: any): Observable<T> {
-        this.defaults = deepMerge(options, Oauth2Service.base);
-
-        let url;
-        let openPopup: Observable<any>;
-        const stateName = this.defaults.name + '_state';
-        const state = this.defaults.state;
-        const exp = new Date(Date.now() + 60 * 60 * 1000).toUTCString();
-        if (typeof state === 'string') {
-            this.storage.set(stateName, state, exp);
-        } else if (typeof state === 'function') {
-            this.storage.set(stateName, state(), exp);
-        }
-
-        url = [this.defaults.authorizationEndpoint, this.buildQueryString()].join('?');
-
-        if (this.config.cordova) {
-            openPopup = this.popup
-                .open(url, this.defaults.name, this.defaults.popupOptions/*, this.defaults.redirectUri*/)
-                .eventListener(this.defaults.redirectUri);
-        } else {
-            openPopup = this.popup
-                .open(url, this.defaults.name, this.defaults.popupOptions/*, this.defaults.redirectUri*/)
-                .pollPopup();
-        }
-
-        return openPopup
-            .pipe(switchMap(oauthData => {
+    open<T extends object | string = any>(oauthOptions: IOauth2Options, userData: object): Observable<T> {
+        const authorizationData = this.getAuthorizationData(oauthOptions);
+        const url = [oauthOptions.authorizationEndpoint, buildQueryString(authorizationData)].join('?');
+        return this.popup.open(url, oauthOptions, this.config.options.cordova).pipe(
+            switchMap((oauthData) => {
                 // when no server URL provided, return popup params as-is.
                 // this is for a scenario when someone wishes to opt out from
                 // satellizer's magic by doing authorization code exchange and
                 // saving a token manually.
-                if (this.defaults.responseType === 'token' || !this.defaults.url) {
+                if (oauthOptions.responseType === 'token' || !oauthOptions.url) {
                     return of(oauthData);
                 }
 
-                if (oauthData.state && oauthData.state !== this.storage.get(stateName)) {
-                    throw 'OAuth "state" mismatch';
+                if (oauthData.state && oauthData.state !== authorizationData.state) {
+                    throw new Error('OAuth "state" mismatch');
                 }
-                let exchangeForToken: any = options.exchangeForToken;
-                if (typeof exchangeForToken !== 'function') {
-                    exchangeForToken = this.exchangeForToken.bind(this);
-                }
-                return exchangeForToken(oauthData, userData);
-            }));
+                return this.exchangeForToken<T>(oauthOptions, authorizationData, oauthData, userData);
+            }),
+        );
     }
 
-    private exchangeForToken(oauthData: { code?: string, state?: string }, userData?: object) {
-        let body = { ...this.defaults, ...oauthData, ...userData };
-
-        let exchangeForTokenUrl = this.config.baseUrl ? joinUrl(this.config.baseUrl, this.defaults.url) : this.defaults.url;
-
-        return this.http.request(
-            this.defaults.method || 'POST',
-            exchangeForTokenUrl, {
-                body,
-                withCredentials: this.config.withCredentials,
-            });
+    private exchangeForToken<T>(options: IOauth2Options, authorizationData: object, oauthData: object, userData: object) {
+        const body = { authorizationData, oauthData, userData };
+        const { baseUrl, withCredentials } = this.config.options;
+        const { url, method = 'POST' } = options;
+        const exchangeForTokenUrl = baseUrl ? joinUrl(baseUrl, url) : url;
+        return this.http.request<T>(method, exchangeForTokenUrl, { body, withCredentials });
     }
 
-    private buildQueryString() {
-        let keyValuePairs: string[][] = [];
-        let urlParams = ['defaultUrlParams', 'requiredUrlParams', 'optionalUrlParams'];
-
-        urlParams.forEach((params) => {
-            if (this.defaults[params]) {
-                (<string[]>this.defaults[params]).forEach((paramName) => {
-                    let camelizedName = camelCase(paramName);
-                    let paramValue = typeof this.defaults[paramName] === 'function' ?
-                        this.defaults[paramName]() :
-                        this.defaults[camelizedName];
-
-                    if (paramName === 'state') {
-                        let stateName = this.defaults.name + '_state';
-                        paramValue = encodeURIComponent(this.storage.get(stateName));
-                    }
-
-                    if (paramName === 'scope' && Array.isArray(paramValue)) {
-                        paramValue = paramValue.join(this.defaults.scopeDelimiter);
-
-                        if (this.defaults.scopePrefix) {
-                            paramValue = [this.defaults.scopePrefix, paramValue].join(this.defaults.scopeDelimiter);
+    private getAuthorizationData(options: IOauth2Options) {
+        const {
+            responseType = 'code',
+            clientId,
+            redirectUri = (window && window.location.origin) || '',
+            scopeDelimiter = ',',
+            scope,
+            state,
+            additionalUrlParams,
+        } = options;
+        const resolvedState = typeof state === 'function' ? state() : state;
+        return [
+            ['response_type', responseType],
+            ['client_id', clientId],
+            ['redirect_uri', redirectUri],
+            ...state ? [['state', resolvedState]] : [],
+            ...scope ? [['scope', scope.join(scopeDelimiter)]] : [],
+            ...additionalUrlParams
+                ? Object
+                    .keys(additionalUrlParams)
+                    .map((key) => {
+                        const value: string | (() => string) | null | undefined =
+                            (additionalUrlParams as any)[key];
+                        if (typeof value === 'string') {
+                            return [key, value];
+                        } else if (typeof value === 'function') {
+                            return [key, value()];
+                        } else if (value === null) {
+                            return [key, ''];
                         }
-                    }
-
-                    if (params !== 'optionalUrlParams' || typeof paramValue !== 'undefined') {
-                        keyValuePairs.push([paramName, paramValue]);
-                    }
-                });
-            }
-        });
-
-        return keyValuePairs.map(function (pair) {
-            return pair.join('=');
-        }).join('&');
+                        return ['', ''];
+                    })
+                : [],
+        ].filter((_) => !!_[0]).reduce((acc, next) => ({ ...acc, [next[0]]: next[1] }), {} as { [key: string]: string });
     }
 }